requirements.txt
playwright>=1.41.0
ratelimit>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
pymongo>=4.6.0
